Move SlashCommandReference to internals
